[chr2img] Add --start and --tile-count
Passing --start will seek to this address in the input file before
reading tiles.  Addresses can be in any format that strconv.ParseInt()
accepts (decimal, hex with 0x prefix, octal with 0 or 0o prefix, and
binary with a 0b prefix).

Passing --tile-count will only read the given number of tiles.  The
accepted number formats are the same as --start.