[extractchr] Add a new command to extract CHR data
This command takes an input ROM, a config, and optionally an output
directory.  The config defines start positions, tile counts, and bit
depths for tiles to be extracted.  Currently, the only accepted bit
depths are 1bpp and 2bpp.  This will change.

Extracted data is written to a set of PNG files in the output directory.