[extractchr] Implement metatile stuff
Implemented metatiles.  Each segment is made up of metatiles.  The
dimensions dictate the number of tiles per metatile.  The count in a
segment is not the number of underlying tiles, but the overall
metatiles.  TileOrder is how the tiles will be ordered in the output
image.

Probably a bunch of cleanup left to do, but this works for now.